Move shouldSkipSerialization to utils module to avoid side-effects of importing from style.js
