+++ packages/block-editor/src/hooks/style.js
@@ -42,6 +42,7 @@ import {
 } from './typography';
 import { SPACING_SUPPORT_KEY, DimensionsPanel } from './dimensions';
 import useDisplayBlockControls from '../components/use-display-block-controls';
+import { shouldSkipSerialization } from './utils';
 
 const styleSupportKeys = [
 	...TYPOGRAPHY_SUPPORT_KEYS,
@@ -161,27 +162,6 @@ function addAttribute( settings ) {
 	}
 
 	return settings;
-}
-
-/**
- * Check whether serialization of specific block support feature or set should
- * be skipped.
- *
- * @param {string|Object} blockType  Block name or block type object.
- * @param {string}        featureSet Name of block support feature set.
- * @param {string}        feature    Name of the individual feature to check.
- *
- * @return {boolean} Whether serialization should occur.
- */
-export function shouldSkipSerialization( blockType, featureSet, feature ) {
-	const support = getBlockSupport( blockType, featureSet );
-	const skipSerialization = support?.__experimentalSkipSerialization;
-
-	if ( Array.isArray( skipSerialization ) ) {
-		return skipSerialization.includes( feature );
-	}
-
-	return skipSerialization;
 }
 
 /**
