+++ packages/block-editor/src/layouts/flow.js
@@ -15,7 +15,7 @@ import { Icon, positionCenter, stretchWide } from '@wordpress/icons';
 import useSetting from '../components/use-setting';
 import { appendSelectors } from './utils';
 import { getGapBoxControlValueFromStyle } from '../hooks/gap';
-import { shouldSkipSerialization } from '../hooks/style';
+import { shouldSkipSerialization } from '../hooks/utils';
 
 export default {
 	name: 'default',
